pyproject.toml
[project]
name = "app-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.4",
    "fastapi>=0.116.1",
    "groq>=0.30.0",
    "ollama>=0.5.1",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.9",
    "uvicorn>=0.35.0",
]
